Move buffer-opening logic into BufferStore
